+++ NEWS.md
@@ -1,3 +1,13 @@
+# multid 0.8.0
+
+* Fixed a typo in D_regularized manual
+* Added difference between two dependent correlations -function (diff_two_dep_cors) which enables simultaneous estimation and testing of Cohen's q under variable dependency
+* Added possibility to use manually constructed regularization and estimation datasets, supplied as a list of two dataframes to "data"-argument in D_regularized
+* Improved the output of diff_two_dep_cors and included an argument for missing data ("ML")
+* Improved clarity in multivariate sex difference vignette
+* Added value_correlation -function for testing and quantifying how ipsatizing values influences associations with other variables
+* Added ddsc_sem -function for deconstructing difference score correlation with structural equation modeling 
+
 # multid 0.7.1
 
 * Added na.rm to qcc bootstrap summary over tau-values